missing cast

+++ co_sim_io/sources/utilities.cpp
@@ -123,7 +123,7 @@ std::unordered_set<std::size_t> ComputePartnerRanksAsImporter(
         }
     } else {
         // several of partner ranks communicate with one rank of me
-        const std::size_t num_ranks_per_partner_rank = std::ceil(PartnerSize / static_cast<double>(MySize));
+        const std::size_t num_ranks_per_partner_rank = static_cast<std::size_t>(std::ceil(PartnerSize / static_cast<double>(MySize)));
         std::unordered_set<std::size_t> partner_ranks;
         const std::size_t lower_end = MyRank*num_ranks_per_partner_rank;
         const std::size_t upper_end = (MyRank+1)*num_ranks_per_partner_rank;
@@ -158,7 +158,7 @@ std::unordered_set<std::size_t> ComputePartnerRanksAsExporter(
     } else {
         if (MySize > PartnerSize) {
             // several of my ranks communicate with one rank of partner
-            const std::size_t num_ranks_per_partner_rank = std::ceil(MySize / static_cast<double>(PartnerSize));
+            const std::size_t num_ranks_per_partner_rank = static_cast<std::size_t>(std::ceil(MySize / static_cast<double>(PartnerSize)));
             return {MyRank/num_ranks_per_partner_rank};
         } else {
             // partner has more ranks, we only communicate with one
